+++ include/RungeKutta.h
@@ -4,9 +4,9 @@ namespace RungeKuttaBenchmark{
       DataSet _dataset;
       vector *_v0;
       unsigned _count_v0;
-      double _h;
+      float _h;
     public:
-      RungeKutta(DataSet dataset, vector *v0, unsigned count_v0, double h); 
+      RungeKutta(DataSet dataset, vector *v0, unsigned count_v0, float h); 
       Fiber *order2();
       Fiber *order4();
   };
