+++ include/rkb_types.h
@@ -1,13 +1,13 @@
 typedef struct vec{
-  double x;
-  double y;
-  double z;
+  float x;
+  float y;
+  float z;
 } vector;
 
 typedef vector *vector_field;
 
 typedef struct tmg{
-  double proc;
-  double memo;
+  float proc;
+  float memo;
   unsigned points_count;
 } timing;
